+++ modules/School_Setup/includes/Modules.inc.php
@@ -248,11 +248,11 @@ function _makeReadMe($module_title,$activated=null)
 	else
 		$module_title_echo = _(str_replace('_', ' ', $module_title));
 
-	//if README file, display in Colorbox
-	if (!isset($_REQUEST['_ROSARIO_PDF']) && file_exists('modules/'.$module_title.'/README'))
+	//if README.md file, display in Colorbox
+	if (!isset($_REQUEST['_ROSARIO_PDF']) && file_exists('modules/'.$module_title.'/README.md'))
 	{
-		//get README content
-		$readme_content = file_get_contents('modules/'.$module_title.'/README');
+		//get README.md content
+		$readme_content = file_get_contents('modules/'.$module_title.'/README.md');
 		
 		// convert MarkDown text to HTML
 		$readme_content = '<div class="markdown-to-html">' . $readme_content . '</div>';