+++ CHANGES.md
@@ -23,7 +23,7 @@ Changes in 2.9
 - Add course_period_school_periods_id to course_period_school_periods table primary key in rosariosis.sql
 - Add Update() function in ProgramFunctions/Update.fnc.php
 - Rework Student Comments: serialize, MarkDown in Students/includes/Comments.inc.php
-- CHANGES, INSTALL & WHATS_NEW files: use MarkDown
+- CHANGES, INSTALL, WHATS_NEW, plugins & modules README files: use MarkDown
 
 Changes in 2.8.11
 -----------------